Update to AFNetworking 2.x, XCTest, Objective-C modules
This makes the project only compatible with iOS 7.x and higher, and OS X 10.9 and higher

+++ ADNKit/ANKClient+ANKHandlerBlocks.h
@@ -17,7 +17,7 @@
 
 typedef void (^AFNetworkingSuccessBlock)(AFHTTPRequestOperation *operation, id responseObject);
 typedef void (^AFNetworkingFailureBlock)(AFHTTPRequestOperation *operation, NSError *error);
-typedef void (^ANKClientFileUploadProgressBlock)(NSUInteger bytesWritten, long long totalBytesWritten, long long totalBytesExpectedToWrite);
+typedef void (^ANKClientFileUploadProgressBlock)(NSUInteger bytesWritten, NSInteger totalBytesWritten, NSInteger totalBytesExpectedToWrite);
 
 
 @interface ANKClient (ANKHandlerBlocks)
